Change struct location to memory for single SLOAD

+++ contracts/modules/capital/CowSwapOperator.sol
@@ -27,7 +27,7 @@ contract CowSwapOperator {
 
   // Constants
   address public constant ETH = 0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE;
-  uint16 public constant MAX_SLIPPAGE_DENOMINATOR = 10000;
+  uint public constant MAX_SLIPPAGE_DENOMINATOR = 10000;
   uint public constant MIN_VALID_TO_PERIOD = 600; // 10 minutes
   uint public constant MAX_VALID_TO_PERIOD = 3600; // 60 minutes
   uint public constant MIN_SELL_AMT_TO_FEE_RATIO = 100; // Sell amount at least 100x fee amount
